requirements.txt
matplotlib==3.7.1
numpy==1.23.5
pandas==1.3.4
pandas_stubs==1.2.0.35
plotly==5.24.1
PyYAML==6.0
scikit_learn==1.1.1
torch==2.5.1
tqdm==4.65.0
